refactor(HIGH-28): 온보딩 페이지에서 ContentSelect 비즈니스 로직 분리 및 props 전달 방식으로 변경
